json_parse: rewrite (around 40% faster)
The json_parse() code has been mostly rewritten. The input data is no
longer written to (it is const char * now). Error reporting is better
(there are two passes on the parser, the first one fast, the second
one slower but checking for error). A system of buffer stacks is used
to keep temporary string/object/array data. There is an overall speedup
of around 40%.