ffglitch: libavutil/quickjs: add FFArray and FFPtr
FFArrays and FFPtrs are based on the built-in Typed Arrays, but with a
lot of functionality and complexity removed to provide a simpler and
faster implementation.

FFArrays are fixed length, which means that once they are created, they
cannot be made larger or smaller.

FFArrays are dense arrays (in contrast to sparse arrays), which means
there are no holes in the array. All elements from 0 to length-1 can be
read/written.