ffglitch: libavutil/quickjs: add MV{,Ref,Array,Ptr,Mask} and MV2D{Array,Ptr,Mask} types
MV and MVRef are a simple representation of a [ horizontal, vertical ]
motion vector.

MVArray and MVPtr are similar to FFArrays and FFPtrs, but they contain
motion vectors instead of simple integer values. There is also a helper
MVMask type.

MV2DArray and MV2DPtr are similar to MVArray and MVPtr, but they
contain a 2-dimensional array of motion vectors instead of a single
array of motion vectors. There is also a helper MV2DMask type.